\name{Voss-package}
\alias{Voss-package}
\alias{Voss}
\docType{package}
\title{ Generic Voss algorithm (random sequential additions) }
\description{
Generating realizations of fractal Brownian functions on uniform 1D & 2D grid with classic and generic versions of the Voss algorithm (random sequential additions).
}
\details{
\tabular{ll}{
Package: \tab Voss\cr
Type: \tab Package\cr
Version: \tab 0.1.5\cr
Date: \tab 2022-05-09\cr
License: \tab GPL-3\cr
}
\code{voss1d()} and \code{voss2d()} functions generate realizations of fractal Brownian functions on uniform 1D & 2D grid with a classic version of the Voss algorithm (random sequential additions).\cr
\code{voss1g()} and \code{voss2g()} functions generate realizations of fractal Brownian functions on uniform 1D & 2D grid with a generic version of the Voss algorithm (random sequential additions).\cr
}
\author{Pavel V. Moskalev}
